Added 'react-spring' library and blob morph
Added the 'react-spring' library to the project to handle the morphing animation for the home page side blobs.

Created a useEffect hook that updates the blob path over a predefined duration.

Switched the Navigation Bar from being in the /App comopnent to being in each of the individual page components. This is intended as I want the animated blob to move behind the navigation bar for the next animation.