Edited Home page to include more animations with 'react=spring' and messed up the Navbar
Added the beginnings of the blob animation, using a seperate blob component that will handle all of the animations that take place with the blobs on the home page. The blob will move position across the screen during while the user scrolls through the home page.

The navigation bar needs to be fixed.